requirements.txt
selenium==4.28.1
webdriver-manager==4.0.2
pandas==2.2.2
google-auth==2.37.0
google-cloud-bigquery==3.29.0
pyarrow==16.1.0
flask==3.0.3
chromedriver-binary==77.0.3865.40.0